requirements.txt
ipykernel
mcp
numpy
pandas
matplotlib
scikit-learn
pydantic
python-dotenv
httpx
fastapi
streamlit
uvicorn
mcp[cli]
langchain
wikipedia
langchain-community
langchain-groq
langchain-openai
langchain-anthropic
python-docx
langchain-core
duckduckgo-search
arxiv